website/requirements.txt
flask==2.0.1
huggingface-hub==0.19.4
pyyaml==6.0.1
docker==6.1.2
pillow==10.0.0
requests==2.31.0
